refactor: replace Redux unread count with TanStack Query, move ReceiveMessagePayload to types
