+++ automotive.marketplace.client/src/features/chat/components/UnreadBadge.tsx
@@ -1,11 +1,12 @@
-import { useAppSelector } from "@/hooks/redux";
-import { selectUnreadCount } from "../state/chatSlice";
+import { useQuery } from "@tanstack/react-query";
+import { getUnreadCountOptions } from "../api/getUnreadCountOptions";
 
 const UnreadBadge = () => {
-  const count = useAppSelector(selectUnreadCount);
+  const { data: unreadQuery } = useQuery(getUnreadCountOptions());
+  const count = unreadQuery?.data.unreadCount ?? 0;
   if (count === 0) return null;
   return (
-    <span className="absolute -top-1 -right-1 flex h-4 w-4 items-center justify-center rounded-full bg-red-500 text-[10px] font-bold text-white">
+    <span className="absolute top-1 -right-1 flex h-4 w-4 items-center justify-center rounded-full bg-red-500 text-[10px] font-bold text-white">
       {count > 99 ? "99+" : count}
     </span>
   );
