+++ automotive.marketplace.client/src/features/chat/constants/chatHub.ts
@@ -3,13 +3,3 @@ export const HUB_METHODS = {
   RECEIVE_MESSAGE: "ReceiveMessage",
   UPDATE_UNREAD_COUNT: "UpdateUnreadCount",
 } as const;
-
-export type ReceiveMessagePayload = {
-  id: string;
-  conversationId: string;
-  senderId: string;
-  senderUsername: string;
-  content: string;
-  sentAt: string;
-  isRead: boolean;
-};
