+++ automotive.marketplace.client/src/features/chat/index.ts
@@ -4,9 +4,5 @@ export { default as MessageThread } from "./components/MessageThread";
 export { default as UnreadBadge } from "./components/UnreadBadge";
 export { useChatHub } from "./api/useChatHub";
 export { useGetOrCreateConversation } from "./api/useGetOrCreateConversation";
-export {
-  default as chatReducer,
-  selectUnreadCount,
-  setUnreadCount,
-} from "./state/chatSlice";
 export type { ConversationSummary } from "./types/ConversationSummary";
+export type { ReceiveMessagePayload } from "./types/ReceiveMessagePayload";
