+++ automotive.marketplace.client/src/features/chat/types/ReceiveMessagePayload.ts
@@ -1,0 +1,9 @@
+export type ReceiveMessagePayload = {
+  id: string;
+  conversationId: string;
+  senderId: string;
+  senderUsername: string;
+  content: string;
+  sentAt: string;
+  isRead: boolean;
+};
