+++ automotive.marketplace.client/src/lib/redux/store.ts
@@ -1,5 +1,4 @@
 import { authReducer } from "@/features/auth";
-import { chatReducer } from "@/features/chat";
 import { configureStore } from "@reduxjs/toolkit";
 import { persistReducer, persistStore } from "redux-persist";
 import storage from "redux-persist/lib/storage";
@@ -15,7 +14,6 @@ const persistedAuthReducer = persistReducer(persistConfig, authReducer);
 export const store = configureStore({
   reducer: {
     auth: persistedAuthReducer,
-    chat: chatReducer,
   },
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
